Stop modules by sending signal to whole process group (kill -pid)
When a module is requested to stop by sending SIGINT or SIGKILL, the signal should also be sent to all subprocesses it might created. This is achieved by passing a negative PID to kill(), which kills all processes in given process group.

This is important when a shell script is used as a module, for example (e.g. to run a Nemea logger and pipe its output to some text-processing tools).